+++ androidHyperskillApp/src/main/java/org/hyperskill/app/android/study_plan/fragment/StudyPlanFragment.kt
@@ -170,6 +170,7 @@ class StudyPlanFragment :
                     }
                 }
             }
+            is StudyPlanScreenFeature.Action.ViewAction.UsersQuestionnaireWidgetViewAction -> TODO()
         }
     }
