+++ shared/src/commonMain/kotlin/org/hyperskill/app/home/injection/HomeComponentImpl.kt
@@ -7,7 +7,6 @@ import org.hyperskill.app.gamification_toolbar.injection.GamificationToolbarComp
 import org.hyperskill.app.home.domain.interactor.HomeInteractor
 import org.hyperskill.app.home.presentation.HomeFeature
 import org.hyperskill.app.interview_preparation.injection.InterviewPreparationWidgetComponent
-import org.hyperskill.app.users_questionnaire.widget.injection.UsersQuestionnaireWidgetComponent
 import ru.nobird.app.presentation.redux.feature.Feature
 
 internal class HomeComponentImpl(private val appGraph: AppGraph) : HomeComponent {
@@ -22,9 +21,6 @@ internal class HomeComponentImpl(private val appGraph: AppGraph) : HomeComponent
 
     private val interviewPreparationWidgetComponent: InterviewPreparationWidgetComponent =
         appGraph.buildInterviewPreparationWidgetComponent()
-
-    private val usersQuestionnaireWidgetComponent: UsersQuestionnaireWidgetComponent =
-        appGraph.buildUsersQuestionnaireWidgetComponent()
 
     override val homeFeature: Feature<HomeFeature.ViewState, HomeFeature.Message, HomeFeature.Action>
         get() = HomeFeatureBuilder.build(
@@ -45,8 +41,6 @@ internal class HomeComponentImpl(private val appGraph: AppGraph) : HomeComponent
             interviewPreparationWidgetComponent.interviewPreparationWidgetReducer,
             interviewPreparationWidgetComponent.interviewPreparationWidgetActionDispatcher,
             interviewPreparationWidgetComponent.interviewPreparationWidgetViewStateMapper,
-            usersQuestionnaireWidgetComponent.usersQuestionnaireWidgetReducer,
-            usersQuestionnaireWidgetComponent.usersQuestionnaireWidgetActionDispatcher,
             appGraph.loggerComponent.logger,
             appGraph.commonComponent.buildKonfig.buildVariant
         )
