+++ shared/src/commonMain/kotlin/org/hyperskill/app/study_plan/screen/presentation/StudyPlanScreenReducer.kt
@@ -9,6 +9,8 @@ import org.hyperskill.app.study_plan.domain.analytic.StudyPlanClickedRetryConten
 import org.hyperskill.app.study_plan.domain.analytic.StudyPlanViewedHyperskillAnalyticEvent
 import org.hyperskill.app.study_plan.widget.presentation.StudyPlanWidgetFeature
 import org.hyperskill.app.study_plan.widget.presentation.StudyPlanWidgetReducer
+import org.hyperskill.app.users_questionnaire.widget.presentation.UsersQuestionnaireWidgetFeature
+import org.hyperskill.app.users_questionnaire.widget.presentation.UsersQuestionnaireWidgetReducer
 import ru.nobird.app.presentation.redux.reducer.StateReducer
 
 internal typealias StudyPlanScreenReducerResult = Pair<StudyPlanScreenFeature.State, Set<StudyPlanScreenFeature.Action>>
@@ -16,6 +18,7 @@ internal typealias StudyPlanScreenReducerResult = Pair<StudyPlanScreenFeature.St
 internal class StudyPlanScreenReducer(
     private val toolbarReducer: GamificationToolbarReducer,
     private val problemsLimitReducer: ProblemsLimitReducer,
+    private val usersQuestionnaireWidgetReducer: UsersQuestionnaireWidgetReducer,
     private val studyPlanWidgetReducer: StudyPlanWidgetReducer
 ) : StateReducer<StudyPlanScreenFeature.State, StudyPlanScreenFeature.Message, StudyPlanScreenFeature.Action> {
     override fun reduce(
@@ -64,6 +67,13 @@ internal class StudyPlanScreenReducer(
                     reduceProblemsLimitMessage(state.problemsLimitState, message.message)
                 state.copy(problemsLimitState = problemsLimitState) to problemsLimitActions
             }
+            is StudyPlanScreenFeature.Message.UsersQuestionnaireWidgetMessage -> {
+                val (usersQuestionnaireWidgetState, usersQuestionnaireWidgetActions) =
+                    reduceUsersQuestionnaireWidgetMessage(state.usersQuestionnaireWidgetState, message.message)
+                state.copy(
+                    usersQuestionnaireWidgetState = usersQuestionnaireWidgetState
+                ) to usersQuestionnaireWidgetActions
+            }
             is StudyPlanScreenFeature.Message.StudyPlanWidgetMessage -> {
                 val (widgetState, widgetActions) =
                     reduceStudyPlanWidgetMessage(state.studyPlanWidgetState, message.message)
@@ -92,6 +102,11 @@ internal class StudyPlanScreenReducer(
                 state.problemsLimitState,
                 ProblemsLimitFeature.Message.Initialize(forceUpdate = retryContentLoadingClicked)
             )
+        val (usersQuestionnaireWidgetState, usersQuestionnaireWidgetActions) =
+            reduceUsersQuestionnaireWidgetMessage(
+                state.usersQuestionnaireWidgetState,
+                UsersQuestionnaireWidgetFeature.InternalMessage.Initialize
+            )
         val (studyPlanState, studyPlanActions) =
             reduceStudyPlanWidgetMessage(
                 state.studyPlanWidgetState,
@@ -108,11 +123,18 @@ internal class StudyPlanScreenReducer(
             emptySet()
         }
 
+        val actions = toolbarActions +
+            problemsLimitActions +
+            usersQuestionnaireWidgetActions +
+            studyPlanActions +
+            analyticActions
+
         return state.copy(
             toolbarState = toolbarState,
             problemsLimitState = problemsLimitState,
+            usersQuestionnaireWidgetState = usersQuestionnaireWidgetState,
             studyPlanWidgetState = studyPlanState
-        ) to (toolbarActions + problemsLimitActions + studyPlanActions + analyticActions)
+        ) to actions
     }
 
     private fun reduceToolbarMessage(
@@ -151,6 +173,26 @@ internal class StudyPlanScreenReducer(
             .toSet()
 
         return problemsLimitState to actions
+    }
+
+    private fun reduceUsersQuestionnaireWidgetMessage(
+        state: UsersQuestionnaireWidgetFeature.State,
+        message: UsersQuestionnaireWidgetFeature.Message
+    ): Pair<UsersQuestionnaireWidgetFeature.State, Set<StudyPlanScreenFeature.Action>> {
+        val (usersQuestionnaireWidgetState, usersQuestionnaireWidgetActions) =
+            usersQuestionnaireWidgetReducer.reduce(state, message)
+
+        val actions = usersQuestionnaireWidgetActions
+            .map {
+                if (it is UsersQuestionnaireWidgetFeature.Action.ViewAction) {
+                    StudyPlanScreenFeature.Action.ViewAction.UsersQuestionnaireWidgetViewAction(it)
+                } else {
+                    StudyPlanScreenFeature.InternalAction.UsersQuestionnaireWidgetAction(it)
+                }
+            }
+            .toSet()
+
+        return usersQuestionnaireWidgetState to actions
     }
 
     private fun reduceStudyPlanWidgetMessage(
